Change property name "dbConn" to "bookRepo"
Indeed, the injected object "dbConn" is used mainly for db transactions. While it is not a singleton for all db transactions across the program, it is only for the Book entity. It makes sense to change its name to "bookRepo", indicating that it's a class for handling db transactions for Book objects.

modified:   backend/src/main/java/com/curtisnewbie/restApi/BookResource.java