Add Order model class
Order class is an Entity class in JPA, wherein the Address class is "embedded" in this class as part of it. Order class should also be related to Book, i.e., an order has one or many books. However, this is not implemented yet.

new file:   backend/src/main/java/com/curtisnewbie/model/Order.java