Create the Data Transfer Object class for Book and Order
As Book and Order classes are used as JPA models that are mapped to the database. They are not serialised to be JSON objects in web services. Thus, I created these two classes as the DTOs for web services. This is an approach that I found to be seamingly useful.

new file:   backend/src/main/java/com/curtisnewbie/dto/BookDTO.java
new file:   backend/src/main/java/com/curtisnewbie/dto/OrderDTO.java