Impl Checkout.service that manages books in busket
Checkout.service is responsible for adding book to and removing book
from busket. Books in busket are items that the user is expecting to
purchase.

This service is injected into book-detail.component and checkout.component.
User in book-detail.component can click the corresponding button to add the
currently displayed book to the busket. In checkout.component, books in busket
are retrieved from this service.