Refactor variable and method name with "busket" to "cart"
"busket" is a typo. It is extensively used in this program in terms of
method name and attribute name. Using the word "cart" is more appropriate
in terms of clarity, e.g., getCart(). For this reason, the methods or
attributes involves using "busket" is now refactored to be using the word
"cart"