Remove two DTO classes
The BookDTO class and OrderDTO class were used as a way to serialise/deserialise
JSON objects. Using DTO was an temporal approach to due with the JPA relations.

Since the Order class do need to serialise the list of books, the DTO for Order
is unnecessary. The List of Order in Book class can be ignored in the processing of
serialisation and deserialisation using @JsonbTransient, thus the BookDTO is also
removed.