Impl CORS for all the HTTP responses
A ContainerResponseFilter is implmeneted to wrap the HTTP responses, wherein
it adds several response headers to enbale CORS. CORS is a security policy
implemented in client side browser. By enabling CORS and setting the origin
manually to localhost:4200, the angualr can now access to the response sent
by this backend REST web services.