Impl fetch-book.service and update related components
The HTTP call to fetch a list of books from backend is implemented. The
book-detail.component and book-list.component are updated to respect to
the received data. E.g., A book may or may not have date, thus, the components
will need to respond to this situation and display "Unknown".