Impl order-confirm component that displays the final order confirmation
Order-confirm component is responsible for displaying the final order
confirmation, where the books in cart and the delivery schemes are
displayed. Before the user can confirm the order (i.e., send the order
to the backend), the user must select one of the delivery option.

Routing for this component and the checkout.component is updated. However
the REST call that sends the order to the backend server is not yet
implmented.