Add address specification in order-confirm.component
In order-confirm.component, users are asked to enter their delivery address.
These inputs are bound to the Delivery object (delivery.ts) using two-way
binding. If the address is not properly specified by the users (i.e, left
empty), the confirm button is not shown. The address is part of the REST
call that sends the order to the backend server.