Add Order interface as a domain model
This class models the JSON object for an order that is sent to the backend
server. For the booksOnOrder, instead of using the Book class, we only stores
the id of each book in the list, as the backend server can use the id as
primary key to find the Book instances.