Setup JPA relationship for the received Order object
Order entity has a many-to-many relationship with Book, where the BookOrder
class is the join table between them. Thus the Order object will have a
List<BookOrder>, wherein, each BookOrder will have the Order object as a
reference that is pointing back to the current Order entity.

However, this Order object is not properly set because of the @JsonbTransient for
serialisation/deserialisation. Here, this is programmatically set.