Change List for cart in CheckoutService to Map
The "cart" in checkout.service was using List data structure to store
the Book that is added into the cart. However, as there can be more than
one instance for each Book, Map is more appropriate than List, where the
key will the unique id of each Book, and the value stores the information
e.g., amount of the Book and the actual Book.

The checkout.component and order-confirm.component are also updated to
use the updated Book object and the Map data structure.