Create resource endpoint to check principal role assigned by JWT authorisation
Microprofile dependency is added as a whole, tho it currently only uses JWT
for authorisation. This resource endpoint is used mainly to facilitate the
development, which tells what the role the request is having. The JWT is a
nice mechanism introduced in microprofile that will be used as a way to
authorise the admin roles (e.g., for POST, PUT and DELETE) in futhre commits.