Change response type of createOrder method in REST endpoints for Order
Previously, this method returns the URI for the created Order instance. However,
client should only be able to review the order belongs to them. Thus, creating
the URI doesn't really make sense. They are the ones that create the order, say
after they successfully make the payment, and returning a response of the created
order back to the client may be a better approach.