Add price property in Order entity
Price propert was not added in the Order entity, as it was considered as a
result that can be automatically calculated as needed. I.e., the price
can be calculated using the list of books that it refers to and its delivery
option. However, as the price of these books can change as well as the
delivery option, this is not appliable.

Now, with the price property added, the price of each order is calculated
right before the order is persisted. And the prices of books and delivery
option that are specified by the client are not trusted. The price of each
order is calculated based on the data in DBMS within a single transaction.