Fix bugs on books.component
Fix the bug that the book.id is not two-way bound to the selectedBook.id,
which causes the failure of POST request when creating a new book. Another
bug is that the id of a existing book is not disabled. Disabling the input
for the id of an existing book is to avoid human error. The condition of
when the three buttons (CREATE, UPDATE and DELETE) are displayed is also
updated for this bug fix.