Change how the selected item is copied
A list of items, e.g., orders are displayed, wherein an admin can select
one to display or to conduct further operations such as UPDATE, DELETE etc.

Previous shallow copy causes the problem, where the displayed/selected item
is changed, while the admin may regret to change it or may not even want to
update it. So, as a solution, instead of copying the reference, we create a
deep copy of the item.