Fix bug: the PUT request for updating Order doesn't setup JPA relation
The JPA is not set when handling PUT request, which results in that
the books in order is not actually added into the order. More seriously,
the books that previously added into the order are also gone because of
the EntityManager.merge() method that update the state entirely.

I.e., it sets the foreign key in the intermediate table that points to this
row in order, to null. This is now fixed by adding the code to setup the JPA
relationship before persistence.