Fix shouldCreateBook() test method
Eventhough previous implementation kinda works, the book returned by
createBook() method should have been the representation of the state
in database, since the EntityManager#persist() method puts it into a
managed state. Such that we don't need to use the getBookById()
method to verify whether the operation is successful or not.